requirements.txt

fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-core
pydantic
langchain_community
chromadb